Rename for consistency.
