+++ Studio/src/Application/Visualization/BarGraph.h
